requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=1.10.7
typing-extensions>=4.7.1
numpy>=1.25.2
torch>=2.2.0
transformers>=4.35.0
beautifulsoup4>=4.12.2
langchain>=0.1.0
langchain-google-genai>=0.0.1
langchain-pinecone>=0.0.1
pinecone-client>=2.2.0
langchain-core>=0.1.0
langchain-community>=0.0.1
langchain-text-splitters>=0.0.1
langgraph>=0.0.1
pdfminer.six
unstructured